Zero page instructions cannot cross a page boundary according to http://www.qmtpro.com/~nes/misc/nestest.txt

+++ gones/cpu6502.go
@@ -270,10 +270,11 @@ func (cpu *CPU) ExecuteInstruction() {
     var operAddr uint16   // address to write/read, if applicable
     var operVal uint8     // value to read
 
+    // http://wiki.nesdev.com/w/index.php/CPU_addressing_modes
     switch instr.AddrMode {
     case Zpg: operAddr = uint16(instr.Operand);                     operPtr = &cpu.Memory[operAddr]
-    case Zpx: operAddr = uint16(instr.Operand) + uint16(cpu.X);     operPtr = &cpu.Memory[operAddr]
-    case Zpy: operAddr = uint16(instr.Operand) + uint16(cpu.Y);     operPtr = &cpu.Memory[operAddr]
+    case Zpx: operAddr = uint16(uint8(instr.Operand) + cpu.X);      operPtr = &cpu.Memory[operAddr]
+    case Zpy: operAddr = uint16(uint8(instr.Operand) + cpu.Y);      operPtr = &cpu.Memory[operAddr]
     case Abs: operAddr = uint16(instr.Operand);                     operPtr = &cpu.Memory[operAddr]
     case Abx: operAddr = uint16(instr.Operand) + uint16(cpu.X);     operPtr = &cpu.Memory[operAddr]
     case Aby: operAddr = uint16(instr.Operand) + uint16(cpu.Y);     operPtr = &cpu.Memory[operAddr]
